// dashboard/src/api/models/run.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */
import type { RunStatus } from './runStatus';
import type { ResultError } from './resultError';
import type { RunArtifactList } from './runArtifactList';
import type { RunNodesList } from './runNodesList';

export interface Run {
  runId: string;
  status: RunStatus;
  definitionHash: string;
  clientId: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  finishedAt?: string | null;
  /** @nullable */
  error?: ResultError;
  /** @nullable */
  artifacts?: RunArtifactList;
  /** @nullable */
  nodes?: RunNodesList;
}