// dashboard/src/api/models/runArtifact.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Scheduler Public API (v1)
 * OpenAPI spec version: 1.3.0
 */
import type { RunArtifactMetadata } from './runArtifactMetadata';

export interface RunArtifact {
  resourceId: string;
  workerId: string;
  type: string;
  /**
   * @minimum 0
   * @nullable
   */
  sizeBytes?: number | null;
  /** @nullable */
  inline?: boolean | null;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  metadata?: RunArtifactMetadata;
}